+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+#### Bugfixes
+
+- [cursor] Fix crash when providing extra large or 0 sizes for cursor theme
+
 ## 0.26.6 -- 2020-05-23
 
 #### Bugfixes